refactor(app): extract the task completion circle into a shared widget
The dashed completion circle and its painter were private to the tasks
page; the home day timeline renders the same control, so pull both into
lib/pages/action_items/widgets/task_completion_circle.dart rather than
copying them.

Verified: flutter analyze lib/pages/action_items -> no new issues;
./test.sh -> 1632 passed.
